cerr_curate_app/requirements.txt
celery==5.2.3
django-celery-beat==2.2.1
Django==3.2.*
django-defender==0.9.1
django-mongoengine==0.4.6
django-rest-framework-mongoengine==3.4.1
django-tz-detect==0.3.0
django-simple-menu==1.2.1
djangorestframework==3.12.*
djangosaml2==1.3.3
drf-yasg==1.20.0
markdown==3.3.3
mongoengine==0.23.0
pymongo==3.11.2
python-dateutil==2.8.1
redis==3.5.3
requests==2.26.*
xmltodict==0.12.0
future==0.18.2
django-mptt
django-crispy-forms
core-main-app
xml_to_dict